+++ src/commands/index.ts
@@ -2,3 +2,4 @@ export * from './accountCommands';
 export * from './dnsCommands';
 export * from './zoneCommands';
 export * from './memberCommands';
+export * from './cacheCommands';